Update files
The fibonacci code needed to show what happened when the user tried to use the fibonacci function with fewer than 2 numbers. This needed to be shown in the output of the file and needed to be included in the zip file. The E-submit receipt was used when handing in the work